Add a host-guest API for VMs without internet requirement (#10)
The guest API forwards API calls to an Aleph node, without the VM needing a network connection.
More functionalities are expected to be added to the API in the future.

Code within a VM can use `aleph-client` directly. Building the runtime requires a version of `aleph-client` without the package `secp256k1`. A branch is currently used for that purpose and NULS1 is temporarily not supported.

The API between the host and VMs now uses `msgpack` for better binary format handling.
Socat is used to expose Vsock on Unix socket.