Fix: Hostname labels must be < 64 chars
Message ids are usually encoded in hexadecimal (base16) across Aleph.im, with a length of 64 characters.

These cannot be used in hostnames since the each label from the hostname (between dots) must be from 1 to 63 characters long.

Since hostnames are case insensitive, base64 cannot be used to encode the hashes.

Solution: Accept base32 encoding of the message ids in the hostname.