Feature: VMs could not be persisted
The supervisor only supported on request execution of
virtual machines.

This add support for an external scheduler to send a
list of VMs that should be started and persisted.

A temporary simple HTTP authentication method is
used to authenticate the scheduler, this is planned to move
to P2P messages from Core Channel Nodes.