Fix: Remove hardcoded 172.0.0.0/8 range from use to comply with RFC1918. Fixes
Previously, 172.0.0.0/8 range was hardcoded to be used as the pool for which to assign IP addresses out of. As this was in violation of RFC1918, this needed to change. As this meant reworking the code anyway, it was changed so the range is configurable. The code to determine the IPs was also moved up to the supervisor level instead of the vm level, as that is more appropriate place for the code to be.