Fix: support running migrations on existing CRNs
Problem: the migration script introduced with alembic assumes
that the records table does not exist yet. This makes it impossible
to run the new version on existing CRNs as the table exists,
making the migration statement fail.

The network_tap column was also deleted in a previous commit
after 0.2.5, but the libsqlite3 version shipped with Ubuntu 20.04
does not support `ALTER TABLE ... DROP COLUMN` statements.

Solution: check if the table exists in the migration script.