Feature: VMs did not support IPv6
Problem: We want each VM to be reachable on the internet, with
no restriction in terms of ports, protocols, etc.

Solution: Slice up the IPv6 range attributed to each CRN and allocate
a subnet to each VM.

Add the support for NDP proxy for some hosting providers.

IPv6 defaults to local addresses for broad compatibility, opt-in public addresses.

The custom init `init1.py` is updated to support IPv6.

Separate loopback networking and always run it
- Remove duplicate IPv4 setting
- Only setup IPv4 route if IPv4 is set
- Only setup eth0 if ip or ipv6

Co-authored-by: Hugo Herter <git@hugoherter.com>