Feature: microVM runtime init version
Problem: the supervisor cannot deduce the version of the init process
of the microVM runtime. This leads to compatibility issues when adding
new features (like IPv6 support in this case).

Solution: define a version for the init process and send this value
to the supervisor when the VM boots. This version is then used to
gatekeep features when configuring the VM.