Fix: Could not SSH into a program VM
Problem: Users and developers could not specify SSH keys to connect inside the VM.

Solution: Add a configuration to VMs that forwards SSH public keys.

Allow developers to specify public keys in the configuration and enable these via a CLI argument.
When developer keys are used, keys from the message are ignored.