Fix: CI for Ubuntu was failing too often because of apt/dpkg locks (#380)
Problem: the CI is looking for only one lock file, but apt/dpkg use several. Furthermore, `lslocks --json` seems to cut its output to a specific width, and configuration options (`--notruncate`) do not appear to have an impact.

Solution: repeat calls to apt-get update until we get the lock. This is not perfect but increases the rate of success.