Implement VM snapshots (#379)
* Problem: Instances loose data between nodes, so we need a way to handle and share this data.

Feature: Implement VM snapshots that will be managed independently.

* Fix: Added some refactor for snapshots and handled some errors.

* Fix: Split snapshot class in two.

* Fix: Make it work well with threads and asyncio.

* Fix: Added code improvements with threading and fixed PR comments.

* Fix: Fixed check_disk method error.

* Fix: Fixed more snapshot errors with multithreading

* Fix: CI for Ubuntu was failing too often because of apt/dpkg locks (#380)

Problem: the CI is looking for only one lock file, but apt/dpkg use several. Furthermore, `lslocks --json` seems to cut its output to a specific width, and configuration options (`--notruncate`) do not appear to have an impact.

Solution: repeat calls to apt-get update until we get the lock. This is not perfect but increases the rate of success.

* Create Ubuntu rootfs (#370)

Updated creation script to create a BTRFS image of Ubuntu filesystem.

* Fix: Use SDK version 0.7.0 (#371)

* Support BTRFS filesystem for instances (#373)

Added support for BTRFS to execute a VM instance.

* Feature: Delete old snapshots after do a new one.

* add dependency to python3-schedule

* Fix latest_snapshot attribute error

* default snapshot frequency 1 hour

---------

Co-authored-by: Olivier Desenfans <desenfans.olivier@gmail.com>
Co-authored-by: Hugo Herter <git@hugoherter.com>
Co-authored-by: Andres D. Molins <nesitor@gmail.com>