Fix: Caddyfile syntax broken in 2.7.4
The syntax of the Caddyfile documented in the quick install guides is broken in Caddy 2.7.4.

This removes the unused `on_demand` section from the configuration, restoring compatibility with newer versions of Caddy.

We maintain the `on_demand_tls` section in case we may re-enable wildcard domains on the default config.