WIP: Add a NodeJS 18 runtime
This starts work in progress to add a runtime with NodeJS 18.

The new runtime based on Debian 12 ships with NodeJS 16, but that is already ancient for the Javascript Hipsters 😅.

This is a copy of the files from the following repository, with differences reduced in `create_disk_image.sh` and `init0.sh` to make the diff more readable.

https://github.com/aleph-im/aleph-github-actions/tree/main/publish-runtime

TODO:
1. Fixes in init1.py still need to be rebased and looked at.
2. An example and tests in CI