Fix: Index did not show IPv6 available
Problem: Node operators could not easily see whether IPv6 is working fine on a node or not.

Solution: We don't want to enforce this yet, so this displays the availability to VMs as additional information with no color indicating a problem, just a black cross.

The connectivity is checked form the diagnostic VM, ensuring that the VM itself has outgoing IPv6 access.