Fix: Prevent concurrent downloads and failed attempts
Problem: The same file could be downloaded multiple times in parallel (ex: same runtime for multiple programs).

Solution: Ensure that only one task is downloading a file at a time by using the temporary file on disk as a lock.

This also adds multiple attempts to download files in case of network issues.