Fix: Connectivity errors crashed the endpoint
Problem:
The endpoint `/status/check/host` should return a data structure with details about the networking that works or does not.

When IPv6 connectivity failed due to a connection error, the endpoint crashed with an error 500 instead of returning the expected value `false` for that type of connectivity.

Solution: Wrap connectivity errors with a try-catch.