Fix: Status check using FastAPI used hardcoded VM ID
Problem: The code was hardcoded to use the same VM ID from the settings for all checks. This did not allow running the same checks on a different VM ID, for example the legacy diagnostic VM with Debian 11, used by existing users.

Solution: Add an argument `vm_id` as optional and default to the value defined in the settings.