Fix: Waiting for another tasks reached timeout
An error was reported on Sentry with the timeout waiting for another task being reached.

The `return` of the function upon success was absent, hence multiple attempts being always retried.

This refactors the download login with 10 attempts to download the file and a 30 seconds timeout waiting for another task.

If the other task ends up failing, this will now take over the download.

The raise of exceptions is also improved for better traceability.