Fix: dictionary changed size during iteration
Iterating over the keys of a dictionary is a lazy
operation that uses a generator under the hood.

An error was reported on Sentry due to
"RuntimeError: dictionary changed size during iteration".

This fixes the issue by transforming the generator
into a list, which is the standard way to approach
this issue in Python.