Problem: error org.freedesktop.systemd1.NoSuchUnit: Unit aleph-vm-controller@...
Jira issue: ALEPH-113

This happened because we tried to check if the service was running
before checking if it was enabled. It resulted in no proper error
but was creating an error message in the logs

Solution: Check if service is enabled before checking if it is active

How to test:
Launch instances. Restart the aleph-vm process, stop them. That error
message should not display