Fix: Error fetching message from vm-connector for start_watch_for_messages_task
Jira ticket : ALEPH-111

Problem:
When aleph-vm was starting if the vm-connector couldn't connect to the
pyaleph API , the startup crashed and systemd restarted it in a loop.

Solution:
Stop registering the sample program at startup, so the  pyaleph api
don't need to be reached at startup. This code was left as a demo but
it didn't really do anything  at the moment.

We have just commented the code if we want to reuse it in the future.

Note:
A more complete and long analysis of the problem, along with propsed
solution is on the JIRA ticket

https://aleph-im.atlassian.net/browse/ALEPH-111?focusedCommentId=10008