Fix: Remove references to slow server api1.aleph.im (#769)
Server api1.aleph.im is very slow and outdated
(Core i7 from 2018, up to 40 seconds to respond
to `/metrics` in the monitoring).

We suspect that this causes issues in the
monitoring and performance of the network.

This branch removes all references to api1 and
replaces them with api3 where relevant.

Co-authored-by: Bram <cortex@worlddomination.be>
Co-authored-by: Olivier Le Thanh Duong <olivier@lethanh.be>