CI: Merge Package workflow and Droplet workflow
This allow to reuse the packages build in the previous workflow
for the droplet test, reducing the number of package build from 12 to 3, in theses workflow.

This:
* fix the issue of package not being able to be built because of rate
limitation on other resources.
* Reduce the chances of random errors.
* Reduce the total CI times requirement.
* Do not attempt to run the droplet test if the package building phase
  fail. (Previously all the package build were launched in parallel
which mean they all failed unecessaryely)
* Make it less costly and faster to run the failed jobs

With theses chance the number of CI failure reduce greatly.
And the cause of failure is more clear