CI: Fix random failures. Merge workflows (#787)
* Make CI a bit more resilient

Do not run export log if cancelled or not setup
Attempt to always do the proper clean up

Print more debug information in case the droplet ipv4 cannot be parsed

* CI: Add timeouts for runtime workflow

* CI: Document the runtime workflow

* CI: Increase timeout to bring up droplet

* CI: pytest. Prevent "Ouput modules" step failure

The dep were not always installed depending on where the failure was
Since this a debug step we don't want it to fail

Do not run it if the workflow was cancelled

* CI: Merge Package workflow and Droplet workflow

This allow to reuse the packages build in the previous workflow
for the droplet test, reducing the number of package build from 12 to 3, in theses workflow.

This:
* fix the issue of package not being able to be built because of rate
limitation on other resources.
* Reduce the chances of random errors.
* Reduce the total CI times requirement.
* Do not attempt to run the droplet test if the package building phase
  fail. (Previously all the package build were launched in parallel
which mean they all failed unecessaryely)
* Make it less costly and faster to run the failed jobs

With theses chance the number of CI failure reduce greatly.
And the cause of failure is more clear

* CI: Wait till an ipv4 on droplet before proceeding

It might be a change in the Digital Ocean API but it return
before the network is setup and with empty setup info
it didn't seem to occur befor so it might be an API change from their
part

* CI: Ensure we reuse the previously calculated IP

* CI: Ensure we use the public IP v4 of the Droplet

Digital Ocean droplet always have a private IP in addition to the public
one.
The API return them in random order so the CI job occasionally tried to
use the internal one and failed.

* CI: Merge the runtime test workflow with the package workflow

Same operation as moving the Droplet workflow, we reuse the already
build package.

The resilience and speed advantage are the sames and add up.

* CI: Remove the deleted workflow, update doc

Rename the main workflow field
Document more

* improve doc

* CI: Do not stop if Export aleph logs command fail

* CI: Print commands run in install aleph step for debug

* CI: Do not fetch whole repository where not needed

In the run_on_droplet job we only require the .github/scripts dir

* CI: Fix apt broken progress bar output