enh: allow user to reboot from inside the VM
JIRA: ALEPH-472

Important note: Previously --no-reboot was used so if the BIOS failed to boot the VM because of for example an invalid disk image provided by the user
then the QEMU process and the controller would halt.
The new behaviour is that the VM will reboot indefinetly. But it should be catched by the wait for init procedure that will test by ping.

This new modification was necessary because the GPU VM need a reboot for their setup.
And generaly user would be confused when rebooted VM would not restart.

The --no-reboot flag was kept for ConfidentialVM as confidential VM do not support reboot anyway