HaProxy: Fix targeting issue
Problem:
Haproxy was not properly configured. And request could go to any server

Analysis:
Previously we added backend server via but not the mapping.
And the map file was not correctly configured

Solution:
Update HaProxy mapping as well as backend server via the control socket
Fix the way mapping is done: Target should be server name not addr:port
Update the haproxy config to only use the fallback server if no mapping if found, otherwise it could be randomly used.
Update tests