Solve Port Forwarding allocation issues (#843)
Fix: I discovered an issue after last Ubuntu updates at the moment of allocating port forwarding rules. When we initialize the firewall ruleset it was creating the chain `aleph-supervisor-prerouting` on the table `raw` instead on the table `nat`. It was raising the issue `Failed to add nftables rules: internal:0:0-0: Error: Could not process rule: Operation not supported` and was leaving the VM without port forwarding features.